hw: pfsoc: can I delete this "alias" stuff
The "ALIAS" sections of memory are not really aliases, at least not in
the way that QEMU currently treats them. Firstly, the size of the
"LO_ALIAS" region, at 0xc000_0000 is not correct and should be 256 MiB,
as there is another region that starts at 0xd000_0000.

As I attempted to describe in a reply to Daniel [0], the addresses that
the CPU writes to get address translated, based on Seg registers.
These Seg registers are runtime configurable, so assuming that specific
sections are aliases of each other doesn't hold water on a generic
basis.
The other potential issue with the aliasing, as is, is that the
"LO_ALIAS" and "HIGH_ALIAS" are non-cached memory. Mapping them strictly
as aliases doesn't reflect how the hardware will behave if you try to
treat those addresses like that!

Realistically I need to go and do some sort of "peripheral" that reads
these values & then maps the memory sections to the correct places at
runtime (if that's even possible) - but if I can't delete the current
(and I believe bogus) aliasing of the non-cached regions to the cached
regions as part of that process, it's not really worth doing.

So, can I delete the "alias" stuff?

Link: https://lore.kernel.org/qemu-devel/Y8ndFjv9prrRXLnL@spud/ [0]
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
---
CC: alistair23@gmail.com
CC: conor@kernel.org
CC: bmeng.cn@gmail.com